requirements.txt
uvicorn
fastapi
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
openai
pytest
pytest-asyncio
stripe
requests
email-validator
PyPDF2
python-docx
python-multipart
pinecone
langchain_openai
langchain_core
langchain_community
tenacity
qrcode[pil]
